requirements.txt
numpy==1.26.4
openai==1.33.0
langchain>0.1.0
opencv_python==4.8.0.76
Pillow==10.3.0
streamlit==1.35.0
